rename library manager file and model
